+++ src/HIndent/Ast/Import/Qualification.hs
@@ -1,0 +1,9 @@
+module HIndent.Ast.Import.Qualification
+  ( Qualification(..)
+  ) where
+
+data Qualification
+  = NotQualified
+  | FullyQualified
+  | QualifiedAs String
+  deriving (Eq)